move files around dont delete originals
