+++ TakeoutMerger.Core/Services/FileService.cs
@@ -1,7 +1,5 @@
 ﻿using System.Collections.Frozen;
-using System.Text.RegularExpressions;
 using Microsoft.Extensions.Logging;
-using TakeoutMerger.Core.Common.Logging;
 using TakeoutMerger.Core.Common.Utils;
 using ZLogger;
