+++ TakeoutMerger.Tests/TakeoutMerger.Tests.csproj
@@ -25,8 +25,4 @@
     <ProjectReference Include="..\TakeoutMerger.Core\TakeoutMerger.Core.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Handlers\" />
-  </ItemGroup>
-
 </Project>
